refactor(world-entry): split chat helpers into world_entry_chat.rs
`world_entry_appearance.rs` grew to 987 lines after the issue-345
chat-helper extraction — over the 700-line hard cap in CLAUDE.md. Move
the chat-channel registration helpers (`DEFAULT_CHAT_CHANNELS`,
`CHAN_TELL`, `build_chat_joined_args`, `build_welcome_message_args`)
and their 6 byte-exact unit tests into a sibling `world_entry_chat.rs`
(231 lines), leaving `world_entry_appearance.rs` at 774 lines — still
over the soft cap but within the hard limit.

The chat-channel logic is also conceptually distinct from appearance:
it implements the `python/base/SGWPlayer.py onClientReady ->
ChannelManager.playerLoggedIn` flow, not the BeingAppearance /
onEntityTint resend handling. Splitting along the natural seam
(per CLAUDE.md file-organization rules) puts the right code under
the right filename.

`handle_on_client_ready` imports from `super::world_entry_chat` and
otherwise unchanged — wire-emit side stays next to the rest of the
`onClientReady` finalisation.

Side effect: tightens the codecov patch-coverage report. With the
helpers + tests scoped to a small file, the per-file coverage view
no longer mixes the 100%-covered helper surface with the
0%-covered async-handler surface — reviewers can see "the chat
helpers are tested" without scrolling past the unreachable
`handle_on_client_ready` body.

Tests: 785/785 (same set, just relocated); fmt + clippy clean.
